webserver: leave no trace!

+++ exfiltrate.py
@@ -44,10 +44,11 @@ class Templates(object):
     def thumbnail(self):
         return "<div style='padding-right: 0.5em; padding-left: 0.5em;'>\
                 <a href='%%URL%%' onclick=\"\
+                document.getElementById('bg').style.display='inline';\
+                document.getElementById('image').style.display='none';\
                 document.getElementById('image').src='%%URL%%'; return false;\
-                \">%%REF%%<br><img src='%%THUMB%%' onerror=\"var that=this;\
-                setTimeout(function(){that.src='%%THUMB%%'}, 10000);\"\
-                width='100%'></a></div><hr>"
+                \">%%REF%%<br><img src='%%THUMB%%' width='100%'></a>\
+                </div><hr>"
     @property
     def html(self):
         return "<!doctype html><html lang='en'><head><meta charset='utf-8'>\
@@ -56,7 +57,12 @@ class Templates(object):
     @property
     def frames_body(self):
         return "<div id='content'><div id='left'>%%THUMBNAILS%%</div>\
-                <div id='right'>%%TITLE%%<br><img id='image' src=''></div></div>"
+                <div id='right'>%%TITLE%%<br>\
+                <img id='bg' style='display:none' src='loading.gif'>\
+                <img id='image' src=''\
+                onload=\"document.getElementById('bg').style.display='none';\
+                this.style.display='inline';\">\
+                </div></div>"
     @property
     def frames_style(self):
         return "<style>html,body {font-size:14pt; height: 100%; overflow: hidden; margin: 0;}\
@@ -64,7 +70,9 @@ class Templates(object):
             "#left {float: left; min-width: 120px; width:8%; height: 100%; overflow: auto; "\
             "box-sizing: border-box;}"\
             "#right {height: 100%; overflow: auto;"\
-            "box-sizing: border-box; padding-left: 0.5em;}</style>"
+            "box-sizing: border-box; padding-left: 0.5em;}"\
+            "img {background: url('loading.gif') no-repeat;}"\
+            "</style>"
 
 Templates = Templates() # properties only work on instances
 
@@ -115,7 +123,6 @@ class Exfiltrator(object):
             return
 
     def generateViewer(self):
-        # Throw in a HTML viewer
         thumbnails = ""            
         for page in range(self._first_page, self._last_page+1):
             p = self._storagedir + "_" + (('%0'+str(self._pad)+"d") % page)
@@ -123,15 +130,14 @@ class Exfiltrator(object):
             thumb = thumb.replace("%%THUMB%%", "thumbs/"+p+"_tnl.jpg")
             thumb = thumb.replace("%%REF%%", "Page "+str(page))
             thumbnails += thumb           
-        with open(os.path.join(self._storagedir, "index.html"), "w") as tf:
-            html = Templates.html
-            html = html.replace("<head>", "<head>" + Templates.frames_style)
-            html = html.replace("%%BODY%%", Templates.frames_body)
-            html = html.replace("%%TITLE%%", self._title)
-            html = html.replace("%%THUMBNAILS%%", thumbnails)
-            tf.write(html)
+        html = Templates.html
+        html = html.replace("<head>", "<head>" + Templates.frames_style)
+        html = html.replace("%%BODY%%", Templates.frames_body)
+        html = html.replace("%%TITLE%%", self._title)
+        html = html.replace("%%THUMBNAILS%%", thumbnails)
+        return html
 
-    def fetch_thumbnail(self, page):
+    def fetch_thumbnail(self, page, no_save=False):
         # These run inside concurrent futures, so check for a terminated main.
         self.exitIfQuit()
         storage = os.path.join(self._storagedir, "thumbs")
@@ -147,17 +153,24 @@ class Exfiltrator(object):
         print("Fetching thumbnail for page "+str(page), end="\r")
         # skip completed thumbnails, but try the actual thumbnail as fallback
         # in case zoom 16 doesn't exist for this page
-        while not os.path.exists(os.path.join(storage, fileName+"_tnl.jpg")):
+        thumbpath = os.path.join(storage, fileName+"_tnl.jpg")
+        while not os.path.exists(thumbpath):
             try:
-                self.fetch_file(url, jp2)
-                subprocess.run(["gm", "mogrify", "-format", "jpg", jp2])
-                os.remove(jp2)
+                if no_save and url.endswith(".jpg"):
+                    return self.fetch_url(url)
+                else:
+                    self.fetch_to_file(url, jp2)
+                    subprocess.run(["gm", "mogrify", "-format", "jpg", jp2])
+                    os.remove(jp2)
+                    if no_save:
+                        f = open(thumbpath, "rb").read()
+                        os.remove(thumbpath)
+                        return f
             except urllib.error.HTTPError as e:
                 if e.code == 404 and url != fallback:
                     url = fallback
                 else:
                     raise
-        return True
 
     def fetch_all_thumbnails(self):
         if not os.path.exists(os.path.join(self._storagedir, "thumbs")):
@@ -166,7 +179,18 @@ class Exfiltrator(object):
         concurrent.futures.wait(pool, None, concurrent.futures.FIRST_EXCEPTION)
         clearline()
 
-    def fetch_file(self, url, to_file):
+    def fetch_url(self, url):
+        while True:
+            try:
+                with urllib.request.urlopen(url, None, 10) as r:
+                    return r.read()
+            except timeout:
+                continue
+
+    def fetch_to_file(self, url, to_file):
+        file_dir = os.path.dirname(to_file)
+        if not os.path.exists(file_dir):
+            os.makedirs(file_dir)
         # don't download twice, but retry on timeouts
         while not os.path.exists(to_file):
             try:
@@ -175,67 +199,71 @@ class Exfiltrator(object):
             except timeout:
                 continue
 
-    def fetch_page(self, page):
+    def fetch_page(self, page, no_save=False):
         curPage = self._document + "_" + (('%0'+str(self._pad)+"d") % page)
         storage = self._storagedir
         fileName = self.boxstr()+curPage
         pageDir = curPage + "_img"
-        
-        # skip completed pages
-        if os.path.exists(os.path.join(storage, fileName+".jpg")):
-            return True
+        pageFile = os.path.join(storage, fileName+".jpg")
 
-        # Pages are composed of sub-image tiles, like a slippy map.
-        # Make a temporary dir for all of the pieces.
-        if not os.path.exists(os.path.join(storage, fileName)):
-            os.makedirs(os.path.join(storage, fileName))
+        # don't download and assemble completed pages
+        if not os.path.exists(pageFile):
+            # Pages are composed of sub-image tiles, like a slippy map.
+            # Make a temporary dir for all of the pieces.
+            if not os.path.exists(os.path.join(storage, fileName)):
+                os.makedirs(os.path.join(storage, fileName))
 
-        # Pages don't have uniform tile counts, so we detect
-        # them for each page starting from something arbitrarily high.
-        # Unfortunately this means not threading all the tile requests.
-        max_y = 50
-        max_x = 50
-        y = 0
-        while y < max_y:
-            x = 0
-            while x < max_x:
-                tileFile = (pageDir + "_TILE_001_" + ('%04d' % y) + "_" 
-                     + ('%04d' % x) + ".JP2")
-                tileURL = self._firstpart + "/" + pageDir + "/" + tileFile
+            # Pages don't have uniform tile counts, so we detect
+            # them for each page starting from something arbitrarily high.
+            # Unfortunately this means not threading all the tile requests.
+            max_y = 50
+            max_x = 50
+            y = 0
+            while y < max_y:
+                x = 0
+                while x < max_x:
+                    tileFile = (pageDir + "_TILE_001_" + ('%04d' % y) + "_" 
+                         + ('%04d' % x) + ".JP2")
+                    tileURL = self._firstpart + "/" + pageDir + "/" + tileFile
 
-                try:
-                    self.fetch_file(tileURL, os.path.join(storage, fileName, tileFile))
-                    x += 1
-                except urllib.error.HTTPError as e:
-                    if e.code == 404:
-                        if y == 0:
-                            max_x = x
+                    try:
+                        tileDest = os.path.join(storage, fileName, tileFile)
+                        self.fetch_to_file(tileURL, tileDest)
+                        x += 1
+                    except urllib.error.HTTPError as e:
+                        if e.code == 404:
+                            if y == 0:
+                                max_x = x
+                            else:
+                                max_y = y
+                                break
                         else:
-                            max_y = y
-                            break
-                    else:
-                        print(str(e))
-                        raise
-            y += 1
+                            print(str(e))
+                            raise
+                y += 1
 
-        clearline()
-        print("Assembling page "+str(page)+".", end="\r")
+            clearline()
+            print("Assembling page "+str(page)+".", end="\r")
 
-        try:
-            # GraphicsMagick Montage is perfect for reassembling the tiles
-            subprocess.run(["gm", "montage", "-mode", "concatenate", "-quality",
-                           "80", "-tile", "%dx%d" % (max_x, max_y),
-                           os.path.join(storage, fileName, "*.JP2"),
-                           os.path.join(storage, fileName+".jpg")])
-        finally:        
-            # Clean up. Erase downloaded tile images for the assembled page.
-            self.cleanup(os.path.join(storage, fileName))
-            if not os.listdir(storage):
-                os.rmdir(storage)
+            try:
+                # GraphicsMagick Montage is perfect for reassembling the tiles
+                subprocess.run(["gm", "montage", "-mode", "concatenate", "-quality", "80", "-tile",
+                                "%dx%d" % (max_x, max_y),
+                                os.path.join(storage, fileName, "*.JP2"),
+                                pageFile])
+            finally:        
+                # Clean up. Erase downloaded tile images for the assembled page.
+                self.cleanup(os.path.join(storage, fileName))
+                if not os.listdir(storage):
+                    os.rmdir(storage)
 
         clearline()
         print("Finished page "+str(page)+".")
-        return True
+        if no_save:
+            f = open(pageFile, "rb").read()
+            os.remove(pageFile)
+            return f
+        return
 
     def fetch_desired_pages(self, start=None, end=None):
         if not start:
@@ -252,16 +280,13 @@ class Exfiltrator(object):
         print("Press Ctrl+C to abort.")
         print("Completed files will be put in the "+self._storagedir+" folder.")
         print("Run again with the same parameters to resume exfiltration.")
-
-        if not os.path.exists(self._storagedir):
-            os.makedirs(self._storagedir)
-
         # Throw in a HTML viewer
         print("Fetching thumbnails")
         self.fetch_all_thumbnails()
         self.exitIfQuit()
         print("Done fetching thumbnails")
-        self.generateViewer()
+        with open(os.path.join(self._storagedir, "index.html"), "w") as tf:
+            tf.write(self.generateViewer())
         self.exitIfQuit()
         print("Fetching pages")
         self.fetch_desired_pages(start, end)
